Refactor get team stream logic
getTeam now checks for default team, and reads team data dinamically with readTeamFile, this way outer services in need of getTeam no longer need to check if a team is in default state or not, abstracting the process and leaving the work to the team storage adapter

+++ backEnd/src/components/Adapters/crestStorage.adapter.ts
@@ -1,27 +1,32 @@
-import { deleteFile, readFile } from "../storage/dataStorage";
-import { getUserCustomCrestFolderPath } from "../storage/userPath";
-import TeamStorageAdapter from "./teamStorage.adapter";
+import { deleteFile, readFile } from '../storage/dataStorage';
+import { getUserCustomCrestFolderPath } from '../storage/userPath';
+import TeamStorageAdapter from './teamStorage.adapter';
 
 const teamStorage = new TeamStorageAdapter();
 export default class CrestStorageAdapter {
-    getCrest(username: string, filename: string): Buffer {
-        const imgPath = `${getUserCustomCrestFolderPath(username)}/${filename}`;
-        const file = readFile(imgPath);
-        return file;
+  getCrest(username: string, filename: string): Buffer {
+    const imgPath = `${getUserCustomCrestFolderPath(username)}/${filename}`;
+    const file = readFile(imgPath);
+    return file;
+  }
+  /**
+   * Deletes old crest if it is different from the new one
+   */
+  deleteOldCrest(
+    username: string,
+    teamId: number | string,
+    newCrestFilename: string,
+  ) {
+    if (teamStorage.isTeamDefault(username, teamId)) return;
+    const team = teamStorage.getTeam(username, teamId);
+    const oldCrestFilename = team.crestUrl.split('/').pop();
+    if (oldCrestFilename !== newCrestFilename) {
+      try {
+        const oldCrestPath = `${getUserCustomCrestFolderPath(username)}/${oldCrestFilename}`;
+        deleteFile(oldCrestPath);
+      } catch (error) {
+        throw error;
+      }
     }
-    /**
-     * Deletes old crest if it is different from the new one
-    */
-    deleteOldCrest(username: string, teamId: number | string, newCrestFilename: string) {
-        const team = teamStorage.getTeam(username, teamId);
-        const oldCrestFilename = team.crestUrl.split('/').pop();
-        if (oldCrestFilename !== newCrestFilename) {
-            try {
-                const oldCrestPath = `${getUserCustomCrestFolderPath(username)}/${oldCrestFilename}`;
-                deleteFile(oldCrestPath);
-            } catch (error) {
-                throw error;
-            }
-        }
-    }
+  }
 }
